Move ast expression definitions to .cpp files
